Clarify use of underscores in cookbook names
The "Use of Hyphens" section previously said:
> Cookbook and custom resource names should contain only alphanumeric characters.

However, the example in the "Cookbook Naming" section includes an underscore,
and [this Discourse answer](https://discourse.chef.io/t/what-is-the-right-style-to-name-a-cookbook/7532/2) says "use underscore-separated, lowercase names".

I've updated the "Use of Hyphens" section to say that underscores are okay as well as alphanumeric characters.